add text in  readme
