+++ README.md
@@ -1,0 +1,34 @@
+# Проект Search_vacancies
+
+Этот проект представляет собой приложение для управления вакансиями, которое позволяет пользователям взаимодействовать с вакансиями через различные функции, такие как добавление, получение, фильтрация и отображение вакансий. Приложение использует абстрактные классы и принципы наследования для обеспечения структурированной и расширяемой кодовой базы.
+## Описание:
+
+- Получение вакансий с API HeadHunter на основе ввода пользователя (модуль interface_by_hh)
+
+- Добавление новых вакансий в локальное хранилище (модуль interface_by_vacancions_json)
+
+- Фильтрация и отображение вакансий по определенным критериям (модуль interface_by_vacancions_json)
+
+- Сортировка вакансий по зарплате (модуль interface_by_vacancions_json)
+
+
+## Установка:
+
+1. Клонируйте репозиторий:
+```
+git clone https://github.com/meirabeloritskaya/search_vacancies.git
+```
+2. Установите зависимости:
+```
+pip install -r requirements.txt
+```
+
+## Документация:
+
+Для получения дополнительной информации обратитесь к [документации](docs/README.md).
+
+## Лицензия:
+
+Этот проект лицензирован по [лицензии MIT](LICENSE).
+
+
